requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.0
attrs==24.3.0
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-extensions==3.2.3
django-organizations==2.5.0
django-phonenumber-field==8.0.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mysqlclient==2.2.5
oauthlib==3.2.2
phonenumbers==8.13.49
pillow==11.0.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
